pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asutp-mixing-module"
version = "0.1.0"
description = "Standalone mini-program for mixing process simulation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
dev = ["pytest>=8.0", "PyYAML>=6.0"]
ui = ["streamlit>=1.40", "pandas>=2.2", "PyYAML>=6.0", "matplotlib>=3.9"]
packaging = ["pyinstaller>=6.0"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
